requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.2.1
dj-database-url==3.0.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
uvicorn==0.34.3
whitenoise==6.9.0
